rename work pool
